Refactor the enum `ArmModeInstruction` to a new file
